feat(modes): implement ACP modes system for per-workflow system prompts
Add comprehensive mode system that allows workflows to define their own
system prompts exposed as ACP modes. This enables agent personas that
are configurable through the standard ACP interface.

Key changes:
- Add Mode struct with support for embedded content and prompt references
- Create 7 new prompt-referencing modes (default, planner, implementer,
  reviewer, tester, committer, rule-checker)
- Add shared prompt partials (_partials/) for coding standards, TDD, git
  practices, and tool use
- Integrate ModeRegistry with llama-agent ACP config
- Add workflow mode support with frontmatter configuration
- Update workflows to reference their corresponding modes
- Clean up deprecated prompt templates

The mode system supports the three-tier override stack:
builtin → user (~/.swissarmyhammer) → project (.swissarmyhammer)

+++ llama-agent/src/types/streaming.rs
@@ -19,7 +19,7 @@ use crate::types::tools::{ToolCall, ToolResult};
 pub struct StreamChunk {
     pub text: String,
     pub is_complete: bool,
-    pub token_count: u32,
+    pub token_count: usize,
     /// Finish reason, only present when is_complete is true
     pub finish_reason: Option<crate::types::generation::FinishReason>,
 }
